Add basic drag-and-drop for plug-ins
Added the initial logic for allowing the user to
drag components from the component catalog and drop
them on the workflow design surface. Note, this is
just preliminary logic in support of an article in
the tutorial series, and will continue to be built
out as the tutorial is written.